requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

python-telegram-bot==20.7
aiohttp==3.9.1

web3==6.0.0
eth-account==0.9.0

pydantic==2.5.0
jinja2==3.1.2
python-multipart==0.0.6
httpx==0.25.2
requests==2.31.0

pandas==2.1.3
numpy==1.24.3
matplotlib==3.8.2

celery==5.3.4
redis==5.0.1

python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2

twilio==8.10.0
